with out adding size in variant edit

+++ kaira/adminapp/views.py
@@ -14,11 +14,10 @@ from django.conf import settings
 import kaira.settings
 from datetime import datetime
 from django.core.mail import send_mail
-from django.contrib.auth.decorators import user_passes_test
 
 
 # Create your views here.
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
+
 def adminlogin(request):
     if request.user.is_superuser:
         return redirect('admindashboard')
@@ -28,14 +27,14 @@ def adminlogin(request):
         user = auth.authenticate(username=username, password=password)
         if user is not None and user.is_staff:
             login(request, user)
-            return redirect('admindashboard')
+            return redirect(admindashboard)
         else:
             messages.error(request, "Username or password mismatch")
     return render(request, 'adminpanel/admin_login.html')
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
-def admindash(request):
-    return render(request, 'adminpanel/admin_dash.html')
+# @user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
+# def admindash(request):
+#     return render(request, 'adminpanel/admin_dash.html')
 
 def adminlogout(request):
     auth.logout(request)
@@ -43,7 +42,6 @@ def adminlogout(request):
             request.session.flush()
     return redirect(adminlogin)
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def add_Category(request):
     if request.method == "POST":
         name = request.POST['name']
@@ -58,12 +56,11 @@ def add_Category(request):
     return render (request,'adminpanel/addcategory.html',locals())
 
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
+
 def view_Category(request):
     category = Category.objects.all()
     return render (request,'adminpanel/viewcategory.html',locals())
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def editcategory(request, pid):
     category = Category.objects.get(id=pid)
     if request.method == 'POST':
@@ -77,7 +74,6 @@ def editcategory(request, pid):
             return redirect(view_Category)
     return render(request, 'adminpanel/editcategory.html', locals())
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def deletecategory(request, pid):
     category = Category.objects.get(id=pid)
     category.delete()
@@ -85,7 +81,7 @@ def deletecategory(request, pid):
     return redirect(view_Category)
 
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
+
 def add_Subcategory(request):
     category = Category.objects.all()
     if request.method == "POST":
@@ -100,17 +96,14 @@ def add_Subcategory(request):
             return redirect(view_Subcategory)
     return render (request,'adminpanel/addsubcategory.html',locals())
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def view_Subcategory(request):
     sub_category = SubCategory.objects.all()
     return render (request,'adminpanel/viewsubcategory.html',locals())
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def view_Brand(request):
     brand = Brand.objects.all()
     return render (request,'adminpanel/viewbrands.html',locals())
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def add_Brand(request):
     if request.method == "POST":
         name = request.POST['name']
@@ -128,7 +121,7 @@ def add_Brand(request):
 #     color = Color.objects.all()
 #     return render (request,'adminpanel/viewcolor.html',locals())
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
+
 def add_Product(request):
     brand=Brand.objects.all()
     sub_category = SubCategory.objects.all()
@@ -149,7 +142,6 @@ def add_Product(request):
         return redirect(view_Product)
     return render (request,'adminpanel/addproduct.html',locals())
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def view_varient(request):
     varient=ProductVarient.objects.all().order_by('id')
     image=ProductImage.objects.all()
@@ -158,8 +150,6 @@ def view_varient(request):
         'image':image,
     }
     return render (request,"adminpanel/viewvarient.html",context)
-
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def view_Product(request):
     product = Product.objects.all().order_by('id')
     
@@ -168,7 +158,6 @@ def view_Product(request):
     }
     return render (request,'adminpanel/viewproduct.html',context)
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def add_varients(request):
     color= ['White','Black','Green','Red','Yellow','Blue','Brown','Orange']
     product=Product.objects.all()
@@ -191,7 +180,7 @@ def add_varients(request):
         Size.objects.create(varient=varient,size=size,stock=stock)
     return render (request,'adminpanel/addproductvarient.html',locals())   
         
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
+
 def edit_Subcategory(request,pid):
     subcategory=SubCategory.objects.get(id=pid)
     category=Category.objects.all()
@@ -208,7 +197,6 @@ def edit_Subcategory(request,pid):
             return redirect('view_Subcategory')
     return render (request,'adminpanel/editsubcategory.html',locals())
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def edit_brand(request,pid):
     brand=Brand.objects.get(id=pid)
     if request.method=="POST":
@@ -222,7 +210,6 @@ def edit_brand(request,pid):
             return redirect('view_Brand')
     return render (request,'adminpanel/editbrand.html',locals())
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def edit_product(request,pid):
     product=Product.objects.get(id=pid)
     category=Category.objects.all()
@@ -247,54 +234,61 @@ def edit_product(request,pid):
             return redirect('view_Brand')
     return render(request,"adminpanel/editproduct.html",locals())
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
+
 def edit_varient(request,pid):
     varient=ProductVarient.objects.get(id=pid)
+    prodsize=Size.objects.filter(varient=varient)
     product=Product.objects.all()
     prodimage=ProductImage.objects.filter(varient=varient)
-    colors=['White','Black','Green','Red','Yellow','Blue','Brown','Orange'] 
+    colors=['White','Black','Green','Red','Yellow','Blue','Brown','Orange']
+    sizes=['XS','S','M','L','XL','XXL','XXXL','7','8','9','10','11','12','28','30','32','34','36','38','40','42','UNI']
     if request.method == "POST":
         try:
-            try:
-                name=request.POST["name"]
-                if ProductVarient.objects.filter(name__iexact=name.lower().replace(' ', '')).exists():
-                    messages.success(request,'category edited successfully')
-                    return redirect('view_varient')
-                else:
-                    varient.name=name
-            except:
-                varient.name=varient.name
-            try:
-                color=request.POST["color"]
-                varient.color=color
-                
-            except:
-                varient.color=varient.color
-            try:
-                prod=request.POST["product"]
-                varient.Product=prod
-            except:
-                varient.Product=varient.Product
-            try:
-                price=request.POST["price"]
-                varient.price=price
-            except:
-                varient.Price=varient.Price
-            try:
-                offer=request.POST["offer"]
-                varient.offer=offer
-            except:
-                varient.offer=varient.offer
-            varient.save()
+            name=request.POST["name"]
+            if ProductVarient.objects.filter(name__iexact=name.lower().replace(' ', '')).exists():
+                messages.success(request,'Varient name  already existing')
+            else:
+                varient.name=name
+        except:
+            varient.name=varient.name
+        try:
+            color=request.POST["color"]
+            varient.color=color
+        except:
+            varient.color=varient.color
+        try:
+            prod=request.POST["product"]
+            varient.Product=prod
+        except:
+            varient.Product=varient.Product
+        try:
+            price=request.POST["price"]
+            varient.price=price
+        except:
+            varient.Price=varient.Price
+        try:
+            offer=request.POST["offer"]
+            varient.offer=offer
+        except:
+            varient.offer=varient.offer
+        varient.save()
+
+        try:
+            l=prodsize.count()
+            for i in range(l):
+                siz=request.POST['size']
+                stock=request.POST['stock']
+                update=Size.objects.get(pk=siz)
+                update.stock=stock
+                update.save()
         except:
             pass
         try:
-
             images = request.FILES.getlist('image')
             hoverimage = request.FILES.get('hover_image')
             # remove the old instances of ProductImage related to the ProductVarient
-            ProductImage.objects.filter(varient=varient).delete()
-            
+            if images and hoverimage is not None:
+                ProductImage.objects.filter(varient=varient).delete()
             for image in images :
                 ProductImage.objects.create(varient=varient, images=image, hover_images=hoverimage)
                 messages.success(request, "Varient images updated successfully.")
@@ -304,7 +298,6 @@ def edit_varient(request,pid):
         
     return render(request,"adminpanel/editvarient.html",locals())
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def delete_category(request,pid):
     try:
         Category.objects.get(id=pid).delete()
@@ -313,7 +306,6 @@ def delete_category(request,pid):
         messages.warning(request,'some thing wrong happend')
     return redirect('view_Category')
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def delete_subcategory(request,pid):
     try:    
         SubCategory.objects.get(id=pid).delete()
@@ -321,8 +313,6 @@ def delete_subcategory(request,pid):
     except:
         messages.warning(request,'some thing wrong happend')
         return redirect(view_Subcategory)
-    
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def delete_brand(request,pid):
     try:
         Brand.objects.get(id=pid).delete()
@@ -331,7 +321,6 @@ def delete_brand(request,pid):
         messages.warning(request,'some thing wrong happend')
     return redirect('view_Brand')
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def delete_product(request,pid):
     try:
         Product.objects.get(id=pid).delete()
@@ -340,7 +329,6 @@ def delete_product(request,pid):
         messages.warning(request,'some thing wrong happend')
     return redirect('view_Product')
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def delete_varient(request,pid):
     try:
         ProductVarient.objects.get(id=pid).delete()
@@ -349,12 +337,10 @@ def delete_varient(request,pid):
         messages.warning(request,'some thing wrong happend')
     return redirect('view_varient')
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def view_coupon(request):
     coupon=Coupon.objects.all()
     return render(request,"adminpanel/viewcoupon.html",locals())
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def add_coupons(request):
     coupon_form = CouponForm(instance=request.user)
     if request.method == 'POST':
@@ -366,7 +352,6 @@ def add_coupons(request):
         form = CouponForm()
     return render(request, 'adminpanel/addcoupon.html', {'coupon_form': coupon_form})
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 def edit_coupon(request, pid):
     coupon = Coupon.objects.get(id=pid)
     coupon_form = CouponForm(instance=request.user)
@@ -382,8 +367,6 @@ def edit_coupon(request, pid):
 
     return render(request, 'adminpanel/editcoupon.html', {'coupon_form': coupon_form, 'coupon': coupon})
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
-
 def delete_coupon(request, pid):
     try:
         Coupon.objects.get(id=pid).delete()
@@ -393,13 +376,9 @@ def delete_coupon(request, pid):
         
     return redirect('view_coupon')
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
-
 def manage_user(request):
     user = User.objects.all().order_by('id')[1:]
     return render(request, 'adminpanel/manageuser.html', {'user': user })
-
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 
 def block_user(request, id):
     user = get_object_or_404(User, id=id)
@@ -412,7 +391,6 @@ def block_user(request, id):
     user.save()
     return redirect('manage_user')
 
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 
 def manage_order(request):
     orders=Order.objects.all().order_by('-id')
@@ -421,9 +399,6 @@ def manage_order(request):
     page_obj = paginator.get_page(page_number) 
     # return render(request, 'manageorder.html', locals())
     return render(request, 'adminpanel/manageorder.html', locals())
-
-
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 
 def update_order(request, id):
     if request.method == 'POST':
@@ -442,8 +417,6 @@ def update_order(request, id):
             except:
                 pass
     return redirect('manage_order')
-
-@user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
 
 def admincancelOrder(request, id):
     
@@ -538,7 +511,6 @@ def admincancelOrder(request, id):
     return redirect(manage_order)
 
 @user_passes_test(lambda u:u.is_superuser,login_url='adminlogin')
-
 def admindashboard(request):
     if request.user.is_authenticated:
         today_sales = Payment.objects.filter(created_at=datetime.today(), paid=True).aggregate(Sum('amount_paid'))['amount_paid__sum']
